requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-crispy-forms==1.7.2
facebook-sdk==3.1.0
idna==2.8
Pillow==6.0.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.2
